Rename main_env_export.py to main_export.py
