+++ conda/cli/main_env.py
@@ -6,13 +6,13 @@ from __future__ import annotations
 from argparse import ArgumentParser, Namespace, _SubParsersAction
 
 from ..deprecations import deprecated
+from . import main_export
 
 
 def configure_parser(sub_parsers: _SubParsersAction | None, **kwargs) -> ArgumentParser:
     from . import (
         main_env_config,
         main_env_create,
-        main_env_export,
         main_env_list,
         main_env_remove,
         main_env_update,
@@ -39,7 +39,7 @@ def configure_parser(sub_parsers: _SubParsersAction | None, **kwargs) -> Argumen
     )
     main_env_config.configure_parser(env_parsers)
     main_env_create.configure_parser(env_parsers)
-    main_env_export.configure_parser(env_parsers)
+    main_export.configure_parser(env_parsers)
     main_env_list.configure_parser(env_parsers)
     main_env_remove.configure_parser(env_parsers)
     main_env_update.configure_parser(env_parsers)
