+++ conda/cli/main_export.py
@@ -23,9 +23,6 @@ def configure_parser(sub_parsers: _SubParsersAction, **kwargs) -> ArgumentParser
         """
         Examples::
 
-            conda env export
-            conda env export --file FILE_NAME
-
             conda export
             conda export --file FILE_NAME
 
@@ -90,7 +87,7 @@ def configure_parser(sub_parsers: _SubParsersAction, **kwargs) -> ArgumentParser
         required=False,
         help="Build environment spec from explicit specs in history",
     )
-    p.set_defaults(func="conda.cli.main_env_export.execute")
+    p.set_defaults(func="conda.cli.main_export.execute")
 
     return p
